refactor: auto index module 분리
