+++ includes/const/html.hpp
@@ -1,17 +1,5 @@
 #include <string.h>
 
-const std::string auto_index_prefix = "<html>\n"
-									  "	<head>\n"
-									  "		<title> Directory Index</title>\n"
-									  "	</head>\n"
-									  "	<body>\n"
-									  "		<h1>Directory Index</h1>"
-									  "	<ul>\n";
-
-const std::string auto_index_suffix = "	</ul>\n"
-									  "	</body>\n"
-									  "</html>\n";
-
 const std::string delete_ok_html = "<html>\n"
 								   "  <body>\n"
 								   "    <h1>File deleted.</h1>\n"
